rename: 폴더구조 세분화
components/profile내 컴포넌트가 많아지면서 조금 더 찾기 쉽게 폴더 구조를 나눔
