--- src/components/Profile/EditModal/ProfileEditModal.module.scss
@@ -1,36 +1,0 @@
-.modal {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 99999;
-  background-color: rgba(0, 0, 0, 0.5);
-  .container {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    background-color: #fff;
-    border-radius: 1rem;
-    padding: 4rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 2rem;
-    .title {
-      margin-right: auto;
-      font-size: 2rem;
-      font-weight: 700;
-    }
-    .buttonContainer {
-      width: 100%;
-      margin-top: 3rem;
-      display: flex;
-      justify-content: space-between;
-    }
-  }
-}
