+++ src/pages/Profile.jsx
@@ -1,7 +1,7 @@
 import { useParams } from 'react-router'
 
 import ProfileInfo from 'components/Profile/Info/ProfileInfo'
-import ProfileItemList from 'components/Profile/ItemList/ProfileItemList'
+import ProfileItemList from 'components/Profile/List/ItemList/ProfileItemList'
 import ListLayout from 'components/Common/Layout/List/ListLayout'
 
 const Profile = () => {
